Split npm package into assemble and deploy
