+++ BUILD
@@ -61,7 +61,8 @@ stardoc(
         "//gcp:lib",
         "//github:lib",
         "//maven:lib",
-        "//npm:lib",
+        "//npm/assemble:lib",
+        "//npm/deploy:lib",
         "//packer:lib",
         "//pip:lib",
         "//rpm:lib",
