+++ npm/assemble/BUILD
@@ -16,3 +16,19 @@
 # specific language governing permissions and limitations
 # under the License.
 #
+
+load("@bazel_skylib//:bzl_library.bzl", "bzl_library")
+
+bzl_library(
+    name = "lib",
+    srcs = [
+        "rules.bzl",
+    ],
+    visibility = ["//visibility:public"]
+)
+
+py_binary(
+    name = "assemble",
+    srcs = ["assemble.py"],
+    visibility = ["//visibility:public"],
+)
