+++ npm/deploy/BUILD
@@ -16,3 +16,34 @@
 # specific language governing permissions and limitations
 # under the License.
 #
+
+load("@bazel_skylib//:bzl_library.bzl", "bzl_library")
+load("@io_bazel_rules_kotlin//kotlin:kotlin.bzl", "kt_jvm_library")
+
+bzl_library(
+    name = "lib",
+    srcs = [
+        "rules.bzl",
+    ],
+    visibility = ["//visibility:public"]
+)
+
+kt_jvm_library(
+    name = "deployer-lib",
+    srcs = glob(["*.kt"]),
+    deps = [
+        "@vaticle_bazel_distribution//common",
+        "@vaticle_bazel_distribution//common/shell",
+        "@vaticle_bazel_distribution//common/util",
+
+        "@maven//:info_picocli_picocli",
+        "@maven//:org_zeroturnaround_zt_exec",
+    ],
+)
+
+java_binary(
+    name = "deployer-bin",
+    runtime_deps = [":deployer-lib"],
+    main_class = "com.vaticle.bazel.distribution.npm.deploy.DeployNPMKt",
+    visibility = ["//visibility:public"],
+)
